+++ src/app/book/list/components/BookGrid.stories.tsx
