+++ src/app/book/list/components/BookItem.stories.tsx
