+++ src/app/book/list/components/BookItem.test.tsx
@@ -1,5 +1,5 @@
 import { fireEvent, render, screen } from '@testing-library/react';
-import BookItem from '@/app/book-list/components/BookItem';
+import BookItem from '@/app/book/list/components/BookItem';
 
 describe('BookItem', () => {
   const setup = (propsOverrides = {}) => {
