+++ src/app/book/list/components/BookItem.tsx
